Add files via upload
Thrilled to discover my very own initiative App for sustainable development. This powerful app empowers users to make eco-friendly choices and contribute to a Greener Planet. This PPT is my initial step of the project. Kudos to my team for creating PPT that aligns with the Global Pursuit for BETTER TOMMOROW!